Update 1.3 [Now Playing]
Added a nowplaying.txt which is generated upon requesting a song. It displays the current Artist - Song info for GPM and the Youtube title, when the track is paused the file is empty. !nowplaying will send the currently playing track (from the file) or announce if the audio is paused.